requirements.txt
torch==1.1.0
torchvision==0.3.0
numpy
# cudatoolkit==10.1
matplotlib
pillow==5.3.0